+++ src/DataTableComponent.php
@@ -139,7 +139,10 @@ abstract class DataTableComponent extends Component
         $this->setupSecondaryHeader();
         $this->setupFooter();
         $this->setupReordering();
-        session([$this->getDataTableFingerprint().'-filters' => $this->{$this->tableName}['filters'] ?? []]);
+        if ($this->getStoreFiltersInSessionIsEnabled())
+        {
+            session([$this->getDataTableFingerprint().'-filters' => $this->{$this->tableName}['filters'] ?? []]);
+        }
 
         return view('livewire-tables::datatable')
             ->with([